Update index.html
HOME PAGE
(Please add this text on the home page)
TwinTownTrivia is available for your private parties, corporate functions or holiday events. We can customize quizzes for you, your team and friends. Click here for rates.
Join us every second Friday (or so) for Brainbusters Online Trivia Quiz – a ZOOM cloud meeting.
The Quizzes start at 8pm start (EDT, UTC– 4 hours).
For details and upcoming events, visit Facebook.com/twintowntrivia72 or refer to the calendar below.
PLEASE REMOVE THE "REGULAR EVENTS" but if you could save those to another doc if/when we want to add that back in